requirements.txt
PySoundCard==0.4.2
PySoundFile==0.4.1
Twisted==13.2.0
autobahn==0.6.5
cffi==0.8.1
docopt==0.6.1
numpy==1.8.0
pycparser==2.10
txpostgres==1.1.0
ujson==1.33
wsgiref==0.1.2
zope.interface==4.0.5
